refactor(webhook): simplify the happy path for invoice.payment_succeeded
